[MOV] hr: separate chat_mixin and employee code
In preparation of the next commit, separate the chat mixin from the
employee js code into appropriate files

Task ID: 2575379
