+++ addons/hr/static/src/js/chat_mixin.js
@@ -1,0 +1,36 @@
+odoo.define('hr.chat_mixin', function (require) {
+"use strict";
+
+const { Component } = owl;
+
+// CHAT MIXIN
+const ChatMixin = {
+    /**
+     * @override
+     */
+    _render: function () {
+        var self = this;
+        return this._super.apply(this, arguments).then(function () {
+            var $chat_button = self.$el.find('.o_employee_chat_btn');
+            $chat_button.off('click').on('click', self._onOpenChat.bind(self));
+        });
+    },
+
+    destroy: function () {
+        if (this.$el) {
+            this.$el.find('.o_employee_chat_btn').off('click');
+        }
+        return this._super();
+    },
+
+    _onOpenChat: function (ev) {
+        ev.preventDefault();
+        ev.stopImmediatePropagation();
+        const env = Component.env;
+        env.messaging.openChat({ employeeId: this.state.data.id });
+        return true;
+    },
+};
+
+return ChatMixin;
+});
